Fix regression due to recursive upbuild support
Recursive upbuild support built in the wrong directory due to changes
that move up a directory when command is 'upbuild'.  The code would
derive a relative path to the build directory (eg '../..') then
derive a second path (for recursive case) by adding ".." to that path.

The problem is that between those two calls we had already changed
directory, so the original path differenc was applied twice.

Added some tests that explicitly test that we run from the expected
directory, as well as a specific test for the recursion case.